[OpenACC] Attach / Detach generic infrastructure
This patch adds the generic infrastructure for the attach and detach
clauses. All of the front ends lower the attach clause as
GOMP_MAP_DETACH data mapping. However, if a detachment is finalized, e.g.

  #pragma acc exit data finalize detach(ptr)

the gimplifier will promote it to GOMP_MAP_FORCE_FINALIZE. Also, this
patch teaches the gimplifier how to ignore GOMP_MAP_STRUCT for the
target update constructs.

Is this patch OK for trunk? I bootstrapped and regression tested it
for x86_64 Linux with nvptx offloading.

Thanks,
Cesar

2018-XX-YY  Cesar Philippidis  <cesar@codesourcery.com>

	gcc/
	* gimplify.c (gimplify_adjust_omp_clauses): Filter out
	GOMP_MAP_STRUCT for acc exit data.
	(gimplify_omp_target_update): Promote GOMP_MAP_DETACH
	to GOMP_MAP_FORCE_DETACH when the finalize clause is present.
	* omp-low.c (lower_omp_target): Add support for GOMP_MAP_{ATTACH,
	DETACH, FORCE_DETACH}.
	* tree-pretty-print.c (dump_omp_clause): Likewise.

	gcc/c-family/
	* c-pragma.h (enum pragma_omp_clause): Define
	PRAGMA_OACC_CLAUSE_{ATTACH,DETACH}.

	include/
	* gomp-constants.h (GOMP_MAP_DEEP_COPY): Define.
	(enum gomp_map_kind): Add GOMP_MAP_{ATTACH, DETACH, FORCE_DETACH}.